make JsonModelConverter internal
