+++ sdk/core/Azure.Core/tests/public/ModelReaderWriterTests/Models/DogListProperty.cs
@@ -150,8 +150,11 @@ namespace Azure.Core.Tests.Public.ModelReaderWriterTests.Models
             {
                 //pulls the additional properties setting from the ModelJsonConverter if it exists
                 //if it does not exist it uses the default value of true for azure sdk use cases
-                var modelConverter = options.Converters.FirstOrDefault(c => c.GetType() == typeof(JsonModelConverter)) as JsonModelConverter;
-                return modelConverter is not null ? modelConverter.Options : ModelReaderWriterHelper.WireOptions;
+                //var modelConverter = options.Converters.FirstOrDefault(c => c.GetType() == typeof(JsonModelConverter)) as JsonModelConverter;
+                //return modelConverter is not null ? modelConverter.Options : ModelReaderWriterHelper.WireOptions;
+
+                //TODO: change the above code once the JsonModelConverter is public
+                return ModelReaderWriterHelper.WireOptions;
             }
         }
         DogListProperty IPersistableModel<DogListProperty>.Create(BinaryData data, ModelReaderWriterOptions options)
