+++ sdk/core/System.ClientModel/tests/ModelReaderWriter/Models/ModelJsonTests.cs
@@ -27,10 +27,5 @@ namespace System.ClientModel.Tests.ModelReaderWriterTests.Models
         [TestCase("W")]
         public void RoundTripWithJsonInterfaceUtf8ReaderNonGeneric(string format)
             => RoundTripTest(format, new JsonInterfaceUtf8ReaderAsObjectStrategy<T>());
-
-        [TestCase("J")]
-        [TestCase("W")]
-        public void RoundTripWithModelJsonConverter(string format)
-           => RoundTripTest(format, new ModelJsonConverterStrategy<T>());
     }
 }
